+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- introduced isolation levels to the fusion strategy
 - added the `one_cell_only` option to `CellInsertionFactory`
 
 ### Changed